Refactor database query wrapper
