+++ lifecycle/lifecycle/database/engine_factory.py
@@ -1,8 +1,8 @@
 import os
 
-from lifecycle.database.engine import DbEngine
-from lifecycle.database.postgres.postgres_engine import PostgresEngine
-from lifecycle.database.sqlite.sqlite_engine import SQLiteEngine
+from lifecycle.database.base_engine import DbEngine
+from lifecycle.database.postgres.engine import PostgresEngine
+from lifecycle.database.sqlite.engine import SQLiteEngine
 from racetrack_client.log.logs import get_logger
 
 logger = get_logger(__name__)
@@ -15,8 +15,8 @@ def create_db_engine() -> DbEngine:
     elif db_type == 'sqlite':
         logger.info('Using local SQLite database')
         return SQLiteEngine()
-    elif db_type == 'sqlite-copy':
-        logger.info('Using copy of local SQLite database')
+    elif db_type == 'sqlite-memory':
+        logger.info('Using in-memory copy of local SQLite database')
         return SQLiteEngine(copy=True)
     else:
         raise ValueError(f'Unknown database type: {db_type}')
